Guillaume branch

+++ Algo_JL_Pure/src/common/message/CounterRequest.java
@@ -5,20 +5,19 @@ import peersim.core.Node;
 /**
  * <p>Represente une requete de compteur.</p>
  */
-public class CounterRequest extends Request{
+public class CounterRequest extends Request {
 
     // Variables.
 
     // Constructors.
 
     /**
-     *
      * @param resourceID
      * @param requestID
      * @param sender
      * @param receiver
      */
-    public CounterRequest(int resourceID, long requestID, Node sender, Node receiver) {
+    public CounterRequest(int resourceID, int requestID, Node sender, Node receiver) {
         super(resourceID, requestID, sender, receiver);
     }
